requirements.txt
boto3==1.34.0
botocore==1.34.0
requests==2.31.0
python-jose[cryptography]==3.3.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opensearch-py==2.4.0
langchain==0.1.0
langchain-aws==0.1.0
numpy==1.24.3
scikit-learn==1.3.0
textblob==0.17.1
transformers==4.36.0
torch==2.1.0
requests-aws4auth==1.3.1
python-dotenv>=1.0.0
